cleanup unused views
